requirements.txt
aiogram==3.21.0
aiohttp==3.12.14
aiofiles==24.1.0
asyncpg==0.30.0
python-dotenv==1.1.1
SQLAlchemy==2.0.41
greenlet==3.2.3

